rename: 사용처에 맞는 디렉토리로 파일이동
