+++ src/function/racingGame.js
@@ -1,9 +1,9 @@
 import { Console } from "@woowacourse/mission-utils";
 import { carsName } from "./carsName.js";
 import { gaemProgress } from "./gameProgress.js";
-import { playCount } from "./playCount.js";
+import { playCount } from "./user/playCount.js";
 import { randomCarsNumber } from "./randomCarsNumber.js";
-import { calculatedResult } from "./CalculatedResult.js";
+import { calculatedResult } from "./util/calculatedResult.js";
 
 export const racingGame = async () => {
   const carsNameArray = await carsName();
