+++ src/function/util/gameProgress.js
